Open Registration for torch::deploy Builtins
Pull Request resolved: https://github.com/pytorch/pytorch/pull/65953

Previously if people want to add a torch::deploy builtin, they need to change torch::deploy internal code (interpreter_impl.cpp) to register the python part as frozen modules and C++ part as builtin modules. This is not convenient and error prone. We want to add open registration support for torch::deploy builtins so that people only need to add one effective line of code in there *library code* to complete the registration.

Here is an example to registry numpy as torch::deploy builtins:
  REGISTER_TORCH_DEPLOY_BUILTIN(numpy, numpy_frozen_modules, PyInit_numpy)

This diff supports open registration of frozen modules. It's the first step to achieve the plan above.
ghstack-source-id: 139487577

Differential Revision: [D31321562](https://our.internmc.facebook.com/intern/diff/D31321562/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D31321562/)!